+++ src/decoy_engine/unmask.py
@@ -42,7 +42,7 @@ import pyarrow as pa
 from decoy_engine.determinism import derive
 from decoy_engine.execution._errors import ExecutionError
 from decoy_engine.execution._strategies._fpe import FPE_KEY_LABEL
-from decoy_engine.plan._compile import _normalize_job_seed
+from decoy_engine.plan._seed import _normalize_job_seed
 from decoy_engine.transforms.fpe import _CHARSETS, fpe_decrypt_value
 
 _LUHN_CAVEAT = (